BotanicaX-functions/requirements.txt
azure-functions>=1.11.0
azure-cosmos>=4.3.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=0.19.0
